:hammer: refact: rename context folder to hooks
